# Add Fabric 1D ring support for all_to_all_dispatch