--- a/tests/ttnn/multidevice_perf_tests/test_all_to_all_dispatch_perf.py
+++ b/tests/ttnn/multidevice_perf_tests/test_all_to_all_dispatch_perf.py
@@ -12,10 +12,10 @@ from models.perf.device_perf_utils import run_device_perf_detailed
 @pytest.mark.parametrize(
     "warmup_iters, arch_type, stage, perf_target_max_us, perf_target_min_us",
     [
-        (5, "6U", "decode", 31, 24),
+        (5, "6U", "decode", 24, 20),
         (5, "TG", "decode", 61, 52),
         (5, "T3K", "decode", 38, 32),  # huge variance among machines for T3K
-        (1, "6U", "prefill", 2000, 1400),
+        (1, "6U", "prefill", 1600, 1300),
         (1, "TG", "prefill", 2700, 2000),
         (1, "T3K", "prefill", 4300, 3800),
     ],