--- a/tests/ttnn/unit_tests/operations/ccl/test_all_to_all_dispatch_6U.py
+++ b/tests/ttnn/unit_tests/operations/ccl/test_all_to_all_dispatch_6U.py
@@ -25,6 +25,7 @@ from tracy import signpost
         {"dispatch_core_axis": ttnn.DispatchCoreAxis.COL, "fabric_config": ttnn.FabricConfig.FABRIC_1D_RING},
         {"dispatch_core_axis": ttnn.DispatchCoreAxis.COL, "fabric_config": ttnn.FabricConfig.FABRIC_2D},
     ],
+    ids=["fabric_1d", "fabric_1d_ring", "fabric_2d"],
     indirect=True,
 )
 @pytest.mark.parametrize("trace_mode", [False])
@@ -114,6 +115,7 @@ def test_all_to_all_dispatch_no_trace(
             "trace_region_size": 500000,
         },
     ],
+    ids=["fabric_1d", "fabric_1d_ring", "fabric_2d"],
     indirect=True,
 )
 @pytest.mark.parametrize("trace_mode", [True])
@@ -194,6 +196,7 @@ def test_all_to_all_dispatch_trace(
             "trace_region_size": 500000,
         }
     ],
+    ids=["fabric_1d_ring"],
     indirect=True,
 )
 @pytest.mark.parametrize("trace_mode", [True])
@@ -273,6 +276,7 @@ def test_decode_perf(
             "trace_region_size": 500000,
         }
     ],
+    ids=["fabric_1d_ring"],
     indirect=True,
 )
 @pytest.mark.parametrize("trace_mode", [True])