--- a/ttnn/cpp/ttnn/operations/ccl/common/host/moe_utils.hpp
+++ b/ttnn/cpp/ttnn/operations/ccl/common/host/moe_utils.hpp
@@ -16,13 +16,6 @@ std::pair<std::vector<tt::tt_metal::IDevice*>, std::array<bool, 4>> get_neighbor
     tt::tt_fabric::Topology topology,
     std::optional<uint32_t> axis);
 
-uint32_t select_link(
-    const distributed::MeshDeviceView& mesh_view,
-    const distributed::MeshCoordinate& src,
-    const distributed::MeshCoordinate& dst,
-    uint32_t num_links,
-    tt::tt_fabric::Topology topology);
-
 // Utilities to code-gen variadic length containers for kernels
 template <typename T>
 std::string stringify(const T& vec) {