### Ticket
#24407

### Problem description
Ring support would add a performance improvement for 1D Fabric Ring topology, so we need it.

### What's changed

- Add some fixes to routing functions
- Delete dead code that affects a2a combine along the way
- TODO: deprecate topology flag in the op as we are moving towards queueing it up directly in the program factory. Since RING is a strict improvement here, there's no need to run the line algorithm on a ring topology.
- Decode B=8/device, S=1 performance: 26.9 us to 22.4 us
- Prefill B=8/device, S=127 performance: 1587 us to 1480 us


### Checklist
- [ ] [All post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml) CI passes https://github.com/tenstorrent/tt-metal/actions/runs/16630511030
- [ ] [Galaxy frequent](https://github.com/tenstorrent/tt-metal/actions/runs/16630507344) CI passes (if applicable)
- [ ] (For runtime and ops writers) [T3000 frequent tests](https://github.com/tenstorrent/tt-metal/actions/runs/16630502492)